requirements.txt
accelerate==0.25.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
cmake==3.27.9
diffusers==0.24.0
distlib==0.3.8
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.19.4
identify==2.5.33
idna==3.6
importlib-metadata==7.0.0
Jinja2==3.1.2
lit==17.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nodeenv==1.8.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pre-commit==3.5.0
psutil==5.9.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
sympy==1.12
tokenizers==0.15.0
torch==2.0.1
tqdm==4.66.1
transformers==4.36.0
triton==2.0.0
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
zipp==3.17.0
